Refactor fqdn.ts
