+++ packages/app/src/cli/services/dev/select-store.ts
@@ -10,9 +10,10 @@ import {error, output, ui, environment} from '@shopify/cli-kit'
 import {partnersRequest} from '@shopify/cli-kit/node/api/partners'
 import {sleep} from '@shopify/cli-kit/node/system'
 import {isSpinEnvironment} from '@shopify/cli-kit/node/environment/spin'
+import {partnersFqdn} from '@shopify/cli-kit/node/environment/fqdn'
 
 const CreateStoreLink = async (orgId: string) => {
-  const url = `https://${await environment.fqdn.partnersFqdn()}/${orgId}/stores/new?store_type=dev_store`
+  const url = `https://${await partnersFqdn()}/${orgId}/stores/new?store_type=dev_store`
   return (
     `Looks like you don't have a dev store in the Partners org you selected. ` +
     `Keep going — create a dev store on Shopify Partners:\n${url}\n`
