+++ packages/cli-kit/src/environment.ts
@@ -1,4 +1,3 @@
 import * as local from './environment/local.js'
-import * as fqdn from './environment/fqdn.js'
 
-export {local, fqdn}
+export {local}
