+++ packages/cli-kit/src/private/node/session.ts
@@ -24,7 +24,7 @@ import * as output from '../../output.js'
 import {firstPartyDev, useDeviceAuth} from '../../environment/local.js'
 import {AbortError} from '../../public/node/error.js'
 import {partnersRequest} from '../../public/node/api/partners.js'
-import {normalizeStoreName, partnersFqdn, identityFqdn} from '../../environment/fqdn.js'
+import {normalizeStoreName, partnersFqdn, identityFqdn} from '../../public/node/environment/fqdn.js'
 import {openURL} from '../../public/node/system.js'
 import {Abort, Bug} from '../../error.js'
 import {gql} from 'graphql-request'
