+++ packages/cli-kit/src/private/node/session/device-authorization.test.ts
@@ -6,13 +6,13 @@ import {
 import {clientId} from './identity.js'
 import {IdentityToken} from './schema.js'
 import {exchangeDeviceCodeForAccessToken} from './exchange.js'
-import {identityFqdn} from '../../../environment/fqdn.js'
+import {identityFqdn} from '../../../public/node/environment/fqdn.js'
 import {shopifyFetch} from '../../../http.js'
 import {err, ok} from '../../../public/node/result.js'
 import {describe, expect, it, vi} from 'vitest'
 import {Response} from 'node-fetch'
 
-vi.mock('../../../environment/fqdn')
+vi.mock('../../../public/node/environment/fqdn.js')
 vi.mock('./identity')
 vi.mock('../../../http')
 vi.mock('./exchange.js')
